refactor: reduce code applying pipe validators in dto

+++ src/endpoints/user/user.controller.ts
@@ -6,23 +6,26 @@ import {
   Param,
   Delete,
   Put,
-  BadRequestException,
+  ParseUUIDPipe,
+  NotFoundException,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdatePasswordDto } from './dto/update-password.dto';
+import { db } from '../../dataBase/db';
+import { ValidateService } from '../../validate/validate.service';
 
 @Controller('user')
 export class UserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private validateService: ValidateService
+  ) {}
 
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
-    if (!createUserDto.login || !createUserDto.password) {
-      throw new BadRequestException(
-        'Required fields (login or password) are not defined'
-      );
-    }
     return this.userService.create(createUserDto);
   }
 
@@ -32,25 +35,30 @@ export class UserController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    if (!this.validateService.doesIdExists(id, db.users)) {
+      throw new NotFoundException("User with this id doesn't exist");
+    }
     return this.userService.findOne(id);
   }
 
   @Put(':id')
   update(
-    @Param('id') id: string,
+    @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updatePasswordDto: UpdatePasswordDto
   ) {
-    if (!updatePasswordDto.newPassword || !updatePasswordDto.oldPassword) {
-      throw new BadRequestException(
-        'Required fields (oldPassword or newPassword) are not defined'
-      );
+    if (!this.validateService.doesIdExists(id, db.users)) {
+      throw new NotFoundException("User with this id doesn't exist");
     }
     return this.userService.updatePassword(id, updatePasswordDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.userService.remove(id);
+  remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    if (!this.validateService.doesIdExists(id, db.users)) {
+      throw new NotFoundException("User with this id doesn't exist");
+    }
+    this.userService.remove(id);
+    throw new HttpException('No content', HttpStatus.NO_CONTENT);
   }
 }
